button to refresh after subscriptions change
